Refactor MixinFogHandler to not use inner class (fixes #72)
